# Improve performance of maybe stack in recursiveTypeRelatedTo