--- a/src/compiler/checker.ts
+++ b/src/compiler/checker.ts
@@ -21303,13 +21303,12 @@ export function createTypeChecker(host: TypeCheckerHost): TypeChecker {
                     if (result === Ternary.True || result === Ternary.Maybe) {
                         // If result is definitely true, record all maybe keys as having succeeded. Also, record Ternary.Maybe
                         // results as having succeeded once we reach depth 0, but never record Ternary.Unknown results.
-                        maybeKeys.popAll(v => {
+                        maybeKeys.popUntilInclusive(id, v => {
                             relation.set(v, RelationComparisonResult.Succeeded | propagatingVarianceFlags);
-                            return v === id;
                         });
                     }
                     else {
-                        maybeKeys.popAll(v => v === id);
+                        maybeKeys.popUntilInclusive(id);
                     }
                 }
                 // Note: it's intentional that we don't pop in the else case;
@@ -21320,7 +21319,7 @@ export function createTypeChecker(host: TypeCheckerHost): TypeChecker {
                 // A false result goes straight into global cache (when something is false under
                 // assumptions it will also be false without assumptions)
                 relation.set(id, (reportErrors ? RelationComparisonResult.Reported : 0) | RelationComparisonResult.Failed | propagatingVarianceFlags);
-                maybeKeys.popAll(v => v === id);
+                maybeKeys.popUntilInclusive(id);
             }
             return result;
         }