For #54939

I did a pprof profile of #54939 and found that a suspicious amount of time was being spent in `recursiveTypeRelatedTo` directly (self time).

![image](https://github.com/microsoft/TypeScript/assets/5341706/9a1695f6-301e-4717-a9fd-b58c6188e610)

The profile didn't really point to anything except for the loops over `maybeKeys`. I noticed that they were all operations where we could be looping over and over again when a set + stack could make every operation constant time.

So, I did that, and the test case from #54939 improved quite a bit:

```
Benchmark 1: old
  Time (mean ± σ):      3.115 s ±  0.067 s    [User: 4.403 s, System: 0.124 s]
  Range (min … max):    3.018 s …  3.196 s    10 runs
 
Benchmark 2: new
  Time (mean ± σ):      2.072 s ±  0.050 s    [User: 3.355 s, System: 0.135 s]
  Range (min … max):    1.985 s …  2.150 s    10 runs
 
Summary
  'new' ran
    1.50 ± 0.05 times faster than 'old'
```

On an even more inflated version of that test, I tripled the number of properties, and:

```
Benchmark 1: old
  Time (mean ± σ):      7.369 s ±  0.178 s    [User: 8.785 s, System: 0.225 s]
  Range (min … max):    7.136 s …  7.685 s    10 runs
 
Benchmark 2: new
  Time (mean ± σ):      3.437 s ±  0.073 s    [User: 4.913 s, System: 0.167 s]
  Range (min … max):    3.363 s …  3.566 s    10 runs
 
Summary
  'new' ran
    2.14 ± 0.07 times faster than 'old'
```

The current version of this PR doesn't appear to regress our regular test suites (some get better).