@jakebailey 2023-08-01
Hm, slower...