@jakebailey 2023-08-01
@typescript-bot perf test this