@jakebailey 2023-08-01
Latest perf test run shows no change / maybe an improvement in xstate, while still preserving the major speedup in the drizzle example. It's rerunning again but that was just a cleanup which should still be as fast.

I'm testing this on some other repos that aren't yet in our test suite e.g. modern mui now.