@jakebailey 2023-08-01
Alright, seems like it's still only neutral or perf positive (or noise). Very cool.

modern mui did not improve from what I can tell, though that's super noisy and hard to test anything but the largest improvements, which this is not.