@ahejlsberg 2023-08-02
I'm certain @weswigham had a PR to this effect some time ago, but now I can't find it.