@jakebailey 2023-08-03
@typescript-bot perf test this faster